+++ src/App.js
@@ -1,13 +1,13 @@
 const express = require('express');
-const authRoutes = require('./vroutes/auth-routes');
-const profileRoutes = require('./vroutes/profile-routes');
-const passportSetup = require('./vconfig/passport-setup');
+const authRoutes = require('../api/vroutes/auth-routes');
+const profileRoutes = require('../api/vroutes/profile-routes');
+const passportSetup = require('../api/vconfig/passport-setup');
 const mongoose = require('mongoose');
-const keys = require('./vconfig/keys');
+const keys = require('../api/vconfig/keys');
 const cookieSession = require('cookie-session');
 const passport = require('passport');
 const bodyParser = require('body-parser');
-const Card = require('./vmodels/card-model.js');
+const Card = require('../api/vmodels/card-model.js');
 const cors = require('cors');
 
 const app = express();
